Refactor Condition
