+++ lib/ransack.rb
@@ -20,7 +20,6 @@ require 'ransack/search'
 require 'ransack/ransacker'
 require 'ransack/translate'
 require 'ransack/adapters/active_record'
-require 'ransack/adapters/active_record/ransack/nodes/condition'
 require 'ransack/context'
 
 ActiveSupport.on_load(:action_controller) do
