+++ lib/ransack/nodes.rb
@@ -3,6 +3,5 @@ require 'ransack/nodes/node'
 require 'ransack/nodes/attribute'
 require 'ransack/nodes/value'
 require 'ransack/nodes/condition'
-require 'ransack/adapters/active_record/ransack/nodes/condition'
 require 'ransack/nodes/sort'
 require 'ransack/nodes/grouping'
